fix(ATE-595): allow Explorer copies onto the redirected Task Files drive (#6)
* fix(ATE-595): report RDP drive capacity without 32-bit overflow

gcsfuse advertises 2^33 allocation units. Storing that count in a 32-bit
int made Windows Explorer see zero free space and refuse copies onto the
redirected drive. Keep the existing RDPDR UINT64 reply and scale huge
counts into a larger block size so ordinary disks are unchanged.


* fix(ATE-595): store RDP drive block counts as 64-bit

Drop the extra scaling step. The RDPDR reply is already UINT64; widening
the in-memory fields is enough to stop gcsfuse's 1 PiB from becoming zero.


---------

+++ src/protocols/rdp/fs.h
@@ -231,13 +231,16 @@ typedef struct guac_rdp_fs_info {
 
     /**
      * The number of free blocks available.
+     *
+     * 64-bit to match RDPDR FileFsSizeInformation and avoid truncating
+     * large backends (for example gcsfuse's 1 PiB) to zero.
      */
-    int blocks_available;
+    uint64_t blocks_available;
 
     /**
      * The number of blocks in the filesystem.
      */
-    int blocks_total;
+    uint64_t blocks_total;
 
     /**
      * The number of bytes per block.
